+++ site/src/main/kotlin/react/bootstrap/site/components/docs/content/typography/Blockquotes.kt
@@ -1,8 +1,8 @@
 package react.bootstrap.site.components.docs.content.typography
 
 import react.RBuilder
-import react.bootstrap.content.typography.blockQuote
-import react.bootstrap.content.typography.blockQuoteFooter
+import react.bootstrap.content.typography.blockquote.blockQuote
+import react.bootstrap.content.typography.blockquote.blockQuoteFooter
 import react.bootstrap.lib.bootstrap.ClassNames
 import react.bootstrap.site.components.docs.fixings.SectionComponent
 import react.bootstrap.site.components.docs.fixings.codeExample
